Making adjustments to Makefile to make it more advanced and efficient.
Started to add macro names for error numbers to make code more readable

Created a new assembly code file that now will add the outfile.in, text
as text into an object file and compile it to be included in the executable
instead of being a file that will be part of a static or dynamic library,
or even a file in a share directory. These methods are ok but then rely
on having those files set in the proper directory and/or library paths
to be edited, etc.

This makes editing the sclex file by hand slightly harder to do after
compilation.

As a quick fix, I didn't really change the code to use the new text
section as a character string but instead, i output the string into a
temporary file, and reread that file as if it existed in the first place.

Then when finished, cleanup should remove the file from the directory.
This isn't as clean as it could be, but works for now.